moved profile to its own file and grocery list button works
