+++ src/main/java/data_access/Constants.java
@@ -4,9 +4,14 @@ package data_access;
  * Constants for the data access layer.
  */
 public class Constants {
-    public static final String API_KEY = "e6a971c78acb49dcbb4e5b324c05d436";
+    // data access
+    public static final String API_KEY = "8c57d019b3ec4934bad61b670ecc45a9";
     public static final String STATUS_CODE_LABEL = "status_code";
     public static final int SUCCESS_CODE = 200;
     public static final String MESSAGE = "message";
+
+    // UI
+    public static final int HEIGHT = 800;
+    public static final int WIDTH = 1200;
 }
