+++ src/main/java/interface_adapter/login/LoginPresenter.java
@@ -24,14 +24,14 @@ public class LoginPresenter implements LoginOutputBoundary {
 
     @Override
     public void prepareSuccessView(LoginOutputData response){
-        // On success, switch to logged in view
-        final LoggedInState loggedInState = new loggedInViewModel.getState();
-        loggedInState.setUsername(response.getUsername());
-        this.loggedInViewModel.setState(loggedInState);
-        this.loggedInViewModel.firePropertyChanged();
-
-        this.viewManagerModel.setState(loggedInViewModel.getViewName());
-        this.viewManagerModel.firePropertyChanged();
+//        // On success, switch to logged in view
+//        final LoggedInState loggedInState = new loggedInViewModel.getState();
+//        loggedInState.setUsername(response.getUsername());
+//        this.loggedInViewModel.setState(loggedInState);
+//        this.loggedInViewModel.firePropertyChanged();
+//
+//        this.viewManagerModel.setState(loggedInViewModel.getViewName());
+//        this.viewManagerModel.firePropertyChanged();
     }
 
     @Override
